+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Language Server: Fixed character names being incorrectly recognised when the colon is not part of the line
 - Flagged the `Utility.AddTagsToLines` method as obsolete.
 - Fixed a bug where escaped characters weren't being correctly added back into the file after adding line tags.
+- declaration statements without a value (`<<declare $var>>`) will no longer crash the compiler.
 
 ### Removed